+++ test/test_helper.rb
@@ -23,3 +23,9 @@ $LOAD_PATH.unshift File.expand_path("../lib", __dir__)
 require "market_hub"
 
 require "minitest/autorun"
+
+# GLOBAL TOKEN VARIABLES
+# Configure valid tokens so that the tests can be carried out successfully.
+@@meli_access_token = 'APP_USR-7241495617113935-012117-2ad6a02378dbdc38d235e308323d5c20-1632856741'
+@@meli_refresh_token = 'TG-65ad893589af6800018e6885-1632856741'
+@@meli_user_id = '1632856741'